requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
future==0.16.0
idna==2.6
python-telegram-bot==8.1.1
requests==2.18.4
urllib3==1.22
